+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+- When calling `Say.say`, `Say.say_item`, and `Say.build_message` without passing in a message String, the "message" result will now be `" ..."`, instead of raising ArgumentError.
+
 - Better definition of the `Say.say_with_block` method:
   1. Rename `footer_message` kwarg to just `footer`
   2. Make `header_message` into a kwarg and rename to `header`